SFTP feature is missing
Related Work Items: #62810

+++ Logitude/General/Usefull Scripts/MetaData/MetaDataScripts.sql
@@ -56,10 +56,10 @@
 ----ObjectFields
 ----alter table advancedqueryfilters drop constraint FK_ObjectFieldAdvancedQueryFilter
 --delete from objectfields where tenant = 0 and (FieldCode not like 'customs.%' or FieldCode is null)
---delete from querycolumns where tenant = 0 and (ObjectFieldCode not like 'customs.%' or ObjectFieldCode is null)
---delete from ScreenFields where tenant = 0 and (ObjectFieldCode not like 'customs.%' or ObjectFieldCode is null)
---delete from AdvancedQueryFilters where tenant = 0 and (ObjectFieldCode not like 'customs.%' or ObjectFieldCode is null)
---delete from RuleConditionFields where tenant = 0  and (ObjectFieldCode not like 'customs.%' or ObjectFieldCode is null)
+--delete from querycolumns where tenant = 0 and (ObjectFieldCode not like 'customs.%' or ObjectFieldCode is null) and userid is null
+--delete from ScreenFields where tenant = 0 and (ObjectFieldCode not like 'customs.%' or ObjectFieldCode is null) 
+--delete from AdvancedQueryFilters where tenant = 0 and (ObjectFieldCode not like 'customs.%' or ObjectFieldCode is null) and userid is null
+--delete from RuleConditionFields where tenant = 0  and (ObjectFieldCode not like 'customs.%' or ObjectFieldCode is null) and systemlevel = 1
 --delete from ObjectTableRuleFields where tenant = 0 and (ObjectFieldCode not like 'customs.%' or ObjectFieldCode is null) and systemlevel = 1
 --delete from ObjectTableRules where tenant = 0 and (TriggerFieldCode not like 'customs.%' or TriggerFieldCode is null) and systemlevel = 1
 
@@ -67,7 +67,7 @@
 --delete from screens where tenant = 0 and (Code not like 'customs.%' or Code is null)
 
 ----Queries
---delete from Queries where tenant = 0 and (Code not like 'customs.%' or Code is null)
+--delete from Queries where tenant = 0 and (Code not like 'customs.%' or Code is null) and userid is null
 
 ----TextCodes
 --delete from MenuButtons where tenant = 0 and (LabelTextCodeCode not like 'customs.%' or LabelTextCodeCode is null)
